对接Sentinel

+++ order-service/src/main/java/com/dili/orders/api/PriceApproveRecordApi.java
@@ -34,7 +34,7 @@ public class PriceApproveRecordApi {
 	public PageOutput<Object> listPageByExample(@RequestBody PriceApproveRecordQueryDto query) {
 		List<PriceApproveRecord> list = this.priceApproveRecordService.listByExample(query);
 		Page<PriceApproveRecord> page = (Page<PriceApproveRecord>) list;
-		return PageOutput.success().setData(page).setPageNum(page.getPageNum()).setTotal((int) page.getTotal());
+		return PageOutput.success().setData(page).setPageNum(page.getPageNum()).setTotal(page.getTotal());
 	}
 
 	/**
